Refactor network layer and add completionHandler..
